ContentAutomator/PostifyNew/requirements.txt
wheel==0.40.0
aiofiles==0.8.0
aiogram==3.0.0b3
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
APScheduler==3.10.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
facebook-sdk==3.1.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
idna==3.4
instagrapi==1.17.10
magic-filter==1.0.9
multidict==6.0.4
outcome==1.2.0
Pillow==9.5.0
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.15.1
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==1.26.6
wsproto==1.2.0
yarl==1.9.2
